Full 560p support + code restructuring
